refactored to allow headless launcher calling the main method of the configured application
